calculate DAS indicator

This function does all preparing steps (selection of relevant plots and reserves) and calculations to obtain the DAS-indicator.

First the relevant forest reserves are selected based on the following criteria:

test for development stages: average diameter (DBH) between 200 and 700 mm

test for desired tree species: > 5 % of the proportion of the basal area should contain trees from the species groups beech, oak, ash/maple or pioneers (birch, willow, esp or rowan)

exclude homogeneous stands: basal area proportion of each tree species group < 98 %

Second, within the forest reserves, relevant plots are selected following the above criteria (so same criteria but different level).

And finally, forest reserves with a total studied area of less than 0.3 ha are removed. Considering A3 plots with an area of 0.0254 ha as the main studied area, forest reserves are removed if less than 12 plots are left.

For the selected forest reserves and plots, the following calculations are made:

arithmetric average of basal area and basal area proportion of the tree species groups for each forest reserve

yearly change of basal area for each forest reserve: \(d = (BA_t1 / BA_t0)^(1 / (t1 - t0)) - 1\) with d = yearly change in %, BA_t1 = average basal area at time 1, BA_t0 = average basal area at time 0 and t1 - t0 = number of years between time 1 and time 0

yearly change of basal area proportion for each forest reserve: \(d = (pBA_t1 / pBA_t0)^(1 / (t1 - t0)) - 1\) with d = yearly change in %, pBA_t1 = average proportional basal area at time 1, BA_t0 = average proportional basal area at time 0 and t1 - t0 = number of years between time 1 and time 0

calculate_DAS_indicator(data_dendro, na.rm = FALSE)

Arguments

data_dendro
dataframe on tree measures with variables plot_id, plottype, tree_measure_id, date_dendro, dbh_mm, height_m, species, alive_dead, decaystage, vol_tot_m3, basal_area_m2, period, OldID, year, subcircle, plotarea_ha,... (output of function load_data_dendrometry())

na.rm
Ignore records with no value? Default is FALSE, so no records will be ignored unless it is explicitly mentioned by 'na.rm = TRUE'

Value

dataframe with results for DAS indicator on forest level (?)

Examples

if (FALSE) {
#change path before running
library(forrescalc)
data_dendro <-
  load_data_dendrometry("C:/MDB_BOSRES_selectieEls/FieldMapData_MDB_BOSRES_selectieEls.accdb")
calculate_DAS_indicator(data_dendro)
}