aggregates parameters on regeneration of trees

This function makes aggregations of tree generation data on the levels of

plot and year (and subplot for core area)

plot, height class and year (and subplot for core area)

plot, tree species and year (and subplot for core area)

plot, height class, tree species and year (and subplot for core area)

For core area plots it makes additional aggregations on the levels of

core area, tree species and year

core area, height class, tree species and year

calculate_regeneration(data_regeneration)

Arguments

data_regeneration

dataframe on tree regeneration with variables plot_id, plottype, subplot_id, height_class, species, nr_of_regeneration, rubbing_damage_number, period, year, subcircle, plotarea_ha, min_number_of_regeneration and max_number_of_regeneration.

Value

List of dataframes that are mentioned in the above description

Examples

library(forrescalc)
# (add path to your own fieldmap database here)
path_to_fieldmapdb <-
  system.file("example/database/mdb_bosres.sqlite", package = "forrescalc")
data_regeneration <- load_data_regeneration(path_to_fieldmapdb)
calculate_regeneration(data_regeneration)
#> $reg_by_plot
#> # A tibble: 30 × 25
#>    plottype plot_id subplot_id period  year number_of_tree_species
#>    <chr>      <int>      <int>  <int> <int>                  <int>
#>  1 CA         11000     155511      1  2000                      0
#>  2 CA         11000     155511      2  2011                      1
#>  3 CA         11000     155511      3  2021                      1
#>  4 CA         11000     155512      1  2000                      0
#>  5 CA         11000     155513      1  2000                      0
#>  6 CA         11000     155521      2  2011                      2
#>  7 CA         11000     155521      3  2021                      1
#>  8 CA         11000     155531      2  2011                      2
#>  9 CA         11000     155531      3  2021                      3
#> 10 CA         21000          1      1  2009                      4
#> # ℹ 20 more rows
#> # ℹ 19 more variables: nr_of_tree_species_established <int>,
#> #   approx_nr_established_ha <dbl>, approx_nr_seedlings_ha <dbl>,
#> #   approx_rubbing_damage_perc_established <dbl>,
#> #   approx_rubbing_damage_perc_seedlings <dbl>,
#> #   rubbing_damage_nr_established_ha <dbl>,
#> #   rubbing_damage_nr_seedlings_ha <dbl>, mean_number_established_ha <dbl>, …
#> 
#> $reg_by_plot_height
#> # A tibble: 76 × 13
#>    plottype plot_id subplot_id period  year height_class number_of_tree_species
#>    <chr>      <int>      <int>  <int> <int>        <int>                  <int>
#>  1 CA         11000     155511      1  2000           NA                      0
#>  2 CA         11000     155511      2  2011         1000                      1
#>  3 CA         11000     155511      2  2011         2000                      1
#>  4 CA         11000     155511      2  2011         3000                      1
#>  5 CA         11000     155511      3  2021         1000                      1
#>  6 CA         11000     155511      3  2021         2000                      1
#>  7 CA         11000     155511      3  2021         3000                      1
#>  8 CA         11000     155511      3  2021         4000                      1
#>  9 CA         11000     155512      1  2000           NA                      0
#> 10 CA         11000     155513      1  2000           NA                      0
#> # ℹ 66 more rows
#> # ℹ 6 more variables: approx_nr_regeneration_ha <dbl>,
#> #   approx_rubbing_damage_perc <dbl>, rubbing_damage_number_ha <dbl>,
#> #   mean_number_of_regeneration_ha <dbl>, lci_number_of_regeneration_ha <dbl>,
#> #   uci_number_of_regeneration_ha <dbl>
#> 
#> $reg_by_plot_species
#> # A tibble: 99 × 24
#>    plottype plot_id subplot_id period  year species approx_nr_established_ha
#>    <chr>      <int>      <int>  <int> <int>   <int>                    <dbl>
#>  1 CA         11000     155511      1  2000      NA                        0
#>  2 CA         11000     155511      2  2011       7                      200
#>  3 CA         11000     155511      3  2021       7                     1000
#>  4 CA         11000     155512      1  2000      NA                        0
#>  5 CA         11000     155513      1  2000      NA                        0
#>  6 CA         11000     155521      2  2011       7                        0
#>  7 CA         11000     155521      2  2011      87                        0
#>  8 CA         11000     155521      3  2021       7                     1700
#>  9 CA         11000     155531      2  2011       7                        0
#> 10 CA         11000     155531      2  2011      87                        0
#> # ℹ 89 more rows
#> # ℹ 17 more variables: approx_nr_seedlings_ha <dbl>,
#> #   approx_rubbing_damage_perc_established <dbl>,
#> #   approx_rubbing_damage_perc_seedlings <dbl>,
#> #   rubbing_damage_nr_established_ha <dbl>,
#> #   rubbing_damage_nr_seedlings_ha <dbl>, mean_number_established_ha <dbl>,
#> #   lci_number_established_ha <dbl>, uci_number_established_ha <dbl>, …
#> 
#> $reg_by_plot_height_species
#> # A tibble: 167 × 13
#>    plottype plot_id subplot_id period  year height_class species
#>    <chr>      <int>      <int>  <int> <int>        <int>   <int>
#>  1 CA         11000     155511      1  2000           NA      NA
#>  2 CA         11000     155511      2  2011         1000       7
#>  3 CA         11000     155511      2  2011         2000       7
#>  4 CA         11000     155511      2  2011         3000       7
#>  5 CA         11000     155511      3  2021         1000       7
#>  6 CA         11000     155511      3  2021         2000       7
#>  7 CA         11000     155511      3  2021         3000       7
#>  8 CA         11000     155511      3  2021         4000       7
#>  9 CA         11000     155512      1  2000           NA      NA
#> 10 CA         11000     155513      1  2000           NA      NA
#> # ℹ 157 more rows
#> # ℹ 6 more variables: approx_nr_regeneration_ha <dbl>,
#> #   approx_rubbing_damage_perc <dbl>, rubbing_damage_number_ha <dbl>,
#> #   mean_number_of_regeneration_ha <dbl>, lci_number_of_regeneration_ha <dbl>,
#> #   uci_number_of_regeneration_ha <dbl>
#> 
#> $reg_by_core_area_species
#> # A tibble: 32 × 25
#>    plottype plot_id period  year species nr_of_subplots_with_regeneration
#>    <chr>      <int>  <int> <int>   <int>                            <int>
#>  1 CA         11000      1  2000      NA                                3
#>  2 CA         11000      2  2011       7                                3
#>  3 CA         11000      2  2011      87                                2
#>  4 CA         11000      3  2021       7                                3
#>  5 CA         11000      3  2021      16                                1
#>  6 CA         11000      3  2021      87                                1
#>  7 CA         21000      1  2009       7                                3
#>  8 CA         21000      1  2009      16                                3
#>  9 CA         21000      1  2009      27                                3
#> 10 CA         21000      1  2009      87                                3
#> # ℹ 22 more rows
#> # ℹ 19 more variables: perc_subplots_with_regeneration <dbl>,
#> #   approx_nr_established_ha <dbl>, approx_nr_seedlings_ha <dbl>,
#> #   approx_rubbing_damage_perc_established <dbl>,
#> #   approx_rubbing_damage_perc_seedlings <dbl>,
#> #   rubbing_damage_nr_established_ha <dbl>,
#> #   rubbing_damage_nr_seedlings_ha <dbl>, mean_number_established_ha <dbl>, …
#> 
#> $reg_by_core_area_height_species
#> # A tibble: 65 × 14
#>    plottype plot_id period  year height_class species nr_of_subplots_with_rege…¹
#>    <chr>      <int>  <int> <int>        <int>   <int>                      <int>
#>  1 CA         11000      1  2000           NA      NA                          3
#>  2 CA         11000      2  2011         1000       7                          3
#>  3 CA         11000      2  2011         1000      87                          2
#>  4 CA         11000      2  2011         2000       7                          2
#>  5 CA         11000      2  2011         3000       7                          1
#>  6 CA         11000      3  2021         1000       7                          3
#>  7 CA         11000      3  2021         1000      16                          1
#>  8 CA         11000      3  2021         1000      87                          1
#>  9 CA         11000      3  2021         2000       7                          3
#> 10 CA         11000      3  2021         3000       7                          3
#> # ℹ 55 more rows
#> # ℹ abbreviated name: ¹​nr_of_subplots_with_regeneration
#> # ℹ 7 more variables: perc_subplots_with_regeneration <dbl>,
#> #   approx_nr_regeneration_ha <dbl>, approx_rubbing_damage_perc <dbl>,
#> #   rubbing_damage_number_ha <dbl>, mean_number_of_regeneration_ha <dbl>,
#> #   lci_number_of_regeneration_ha <dbl>, uci_number_of_regeneration_ha <dbl>
#>