retrieve data on shoots from Fieldmap database

This function queries the given database to retrieve additional data on shoots to use with dendrometry data.

load_data_shoots(database, extra_variables = FALSE)

Arguments

database

name of Fieldmap/Access database (with specific Fieldmap structure) including path

extra_variables

Should additional variables such as iufro_hght, iufro_vital, iufro_socia, remark and common_remark be added? Default is FALSE (no).

Value

Dataframe with shoot data

Examples

library(forrescalc)
# (add path to your own fieldmap database here)
path_to_fieldmapdb <-
  system.file("example/database/mdb_bosres.sqlite", package = "forrescalc")
load_data_shoots(path_to_fieldmapdb)
#>    plot_id period tree_measure_id shoot_measure_id dbh_mm height_m intact_snag
#> 1      101      1           11557                1     60     7.00          11
#> 2      101      1           11557                2     60       NA          11
#> 3      101      1           11599                1     60     7.00          11
#> 4      101      1           11599                2     60       NA          11
#> 5      101      1           11597                1    130    14.00          11
#> 6      101      1           11597                2    130       NA          11
#> 7      101      1           11597                3    100       NA          11
#> 8      101      1           11598                1     90    11.00          11
#> 9      101      1           11609                1     90    10.00          11
#> 10   21000      1              55                1     95       NA          11
#> 11   21000      1              52                1    440    29.33          11
#> 12   21000      1              52                2    375       NA          11
#> 13   21000      1              52                3    335       NA          11
#> 14   21000      1              61                1    120    18.27          11
#> 15   21000      1              61                2    110       NA          11
#> 16   21000      1              36                1    355    21.50          11
#> 17   21000      1              36                2     80       NA          11
#> 18   21000      1              36                3     60       NA          11
#> 19  141100      1             323                1     50     4.00          11
#> 20     101      2           11557                1     60       NA          11
#> 21     101      2           11557                2     50       NA          11
#> 22     101      2           11599                1     60       NA          11
#> 23     101      2           11599                2     60       NA          11
#> 24     101      2           11597                1    145       NA          11
#> 25     101      2           11597                2    140       NA          11
#> 26     101      2           11597                3    110       NA          11
#> 27     101      2           11598                1    115       NA          11
#> 28     101      2           11609                1     90       NA          11
#> 29   21000      2              26                1     60       NA          11
#> 30   21000      2              20                1    440    30.50          11
#> 31   21000      2              20                2    335       NA          11
#> 32   21000      2              20                3    480       NA          11
#> 33   21000      2              80                1    130       NA          11
#> 34   21000      2              80                2    120       NA          11
#> 35   21000      2             154                1    370    24.50          11
#> 36   21000      2             154                2     85       NA          11
#> 37  141100      2              64                1     65       NA          11
#> 38     101      3              13                1     50       NA          11
#> 39     101      3              12                1     60       NA          11
#> 40     101      3              12                2     55       NA          11
#> 41     101      3               8                1    145       NA          11
#> 42     101      3               8                2    150       NA          11
#> 43     101      3               8                3    115       NA          11
#> 44     101      3               5                1     55       NA          11
#>    decaystage
#> 1          16
#> 2          16
#> 3          16
#> 4          16
#> 5          16
#> 6          16
#> 7          16
#> 8          16
#> 9          16
#> 10         12
#> 11         16
#> 12         16
#> 13         16
#> 14         16
#> 15         16
#> 16         16
#> 17         16
#> 18         16
#> 19         16
#> 20         12
#> 21         12
#> 22         12
#> 23         12
#> 24         16
#> 25         16
#> 26         16
#> 27         13
#> 28         12
#> 29         12
#> 30         16
#> 31         16
#> 32         16
#> 33         16
#> 34         16
#> 35         16
#> 36         16
#> 37         16
#> 38         NA
#> 39         16
#> 40         16
#> 41         NA
#> 42         NA
#> 43         NA
#> 44         16